+++ README.md
@@ -8,9 +8,9 @@ This is a map of sustainable businesses in Bergen. It is made with Astro and Map
 
 The pins on the map come from the Overpass API, which is an API that retrieves data about features or POIs (points of interest) mapped by OpenStreetMap volunteers. This means that in order to update something on our map, you must update
 
-The query we use to retrieve data can be found in [`src/data/query.overpassql`](src/data/query.overpassql).
+The queries we use to retrieve data can be found in [`src/overpass/queries`](src/overpass/queries).
 
-You can try it out in the [Overpass API sandbox](https://overpass-turbo.eu/).
+You can try them out in the [Overpass turbo sandbox](https://overpass-turbo.eu/) by copy pasting the entire contents of an .overpassql file.
 
 ## 🚀 Project Structure
