+++ src/components/KindBadge.tsx
