+++ src/components/Panel.tsx
@@ -1,9 +1,9 @@
 import { useStore } from "@nanostores/react";
 import { $feature, $showInfoPanel, hideInfoPanel } from "../store/feature";
 import "./panel.css";
-import { OpeningHours } from "./opening-hours";
+import { OpeningHours } from "./OpeningHours";
 import { formatAddress } from "../utils/format-address";
-import KindBadge from "./kind-badge";
+import KindBadge from "./KindBadge";
 
 export const Panel = () => {
   const show = useStore($showInfoPanel);
